Move FeatureFlag provider into Duck
